tried to separate code, problems with import statements
